Spit off some pltform-specific code to make linter happy
